requirements.txt
accelerate==0.14.0
click==8.1.3
Flask==2.2.2
Flask-Cors==3.0.10
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.4
packaging==21.3
psutil==5.9.4
pyparsing==3.0.9
python-dotenv==0.21.0
six==1.16.0
torch==1.13.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
diffusers==0.7.2
filelock==3.8.0
huggingface-hub==0.11.0
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.4
packaging==21.3
Pillow==9.3.0
pyparsing==3.0.9
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
tokenizers==0.13.2
tqdm==4.64.1
transformers==4.24.0
typing-extensions==4.4.0
urllib3==1.26.12
Werkzeug==2.2.2
zipp==3.10.0
